pyproject.toml
[project]
name = "cocoindex-etl-with-document-ai"
version = "0.1.0"
description = "Extract and embed PDF documents using Google Document AI and Cocoindex."
requires-python = ">=3.10"
dependencies = [
    "cocoindex>=0.1.20",
    "python-dotenv>=1.0.1",
    "google-cloud-documentai>=2.20.1",
    "psycopg-pool>=0.1.0"
]
